Write rake task to update submission scores after meetings finish
Rake task to passback grades of all active resources' submissions.

Considering we just want to run this every day until we release the
updated app in Spring, I figured it doesn't need to be as deliberate
about what submissions we run this on.

Round compute score so comparison in #update_score works.

The score comparison within #update_score was bugged because the score
and compute_score returned floats of different precision.

Nifty rubocop fix.