Update student show page to use Attendances
Link to FontAwesome pro license

Standardize width/height of icon column

Display attendance status available actions

wip

Add checked_in_at for manual approval

Remove Check in button via CheckIn model

Switch to Bootstrap gem.

I couldn't get the media-breakpoint-up mixin to work in `meetings.scss`.
This might be trying to hammer a nail with a sledgehammer but it made it
work.

Display status in copy

DRY

Customize size of icons for multitude of viewports

wip

Enable Bullet

Rearrange buttons on card; use helper

The html for the buttons for "Check in" and "Appeal" was getting huge so
I made a helper. I don't feel like any of this is best practice... I
probably should be rendering a partial instead of returning raw html but
for now it's working.

Add IDs to meeting cards

Remove side "Check in" button

Ultimately, we're hoping to display Checkins in an iframe within Canvas
and in doing so, we like the disign of content being in a single column.

Also, considering each meeting now has large buttons to check in and
appeal, the side "Check in" button loses attention and is superfluous.

Write helper to build single button

The build_attendance_event_buttons helper method created multiple
buttons. I decided it would be better to have a method that would build
a single thing and then call it as many times as needed.

attendance_event_button now builds a single button when given an
Attendance and Event as arguments.

I've implemented CSS grid for the buttons. Each button built by the
helper method will span a single column and row in a CSS grid. This
means care will have to be taken to call the helper method inside of a
div that has grid-template CSS assigned to it with the appropriate
structure to satisfy the buttons you want to build.

Remove unused helper

Changes suggested by Hound

Simplify copy

Extraneous copy was forcing new lines

Remove explicit script link for jquery

Disable Rollbar for Bullet for now

Swap bootstrap-sprockets for bootstrap