Add caching to Attendances for performance
Add meta_request gem

Use variants for resource show

Add caching to attendances on teacher page

Move variant code into show action

When navigating to the backdoor routes, it's likely that no
`current_enrollment` will be stored. We can't do the variant conditional
in a before_action because of this.