Make student attendances collapsible. Resolves #66
This lets students check in to future meetings.

Make meetings collapsible

Future meetings are collapsed by default but a student can click it to
display the event buttons now.

Show attendance if it's not pending

Upon clicking an event button for a future meeting, AJAX would replace
the element with the partial. Since the meeting was still in the future,
the replacement partial would be hidden. This shows the partial if the
meeting has occurred OR if action has been taken on it (it is not
pending)

Shorten line lengths.